3 - characters: clean-up character name component
