+++ .eslintrc.js
@@ -149,8 +149,6 @@ module.exports = {
     viewportModes: 'writable',
     teleportUserInLevel: 'writable',
     sendEvent: 'writable',
-    characterNameColors: 'writable',
-    CharacterNameText: 'writable',
     CharacterReaction: 'writable',
     sendAudioChunksToNearUsers: 'writable',
     sendAudioChunksToUsersInZone: 'writable',
